Explicit cast to help g++
G++ is not smart enough to pick the correct function specialization.
Let's help him a little bit.

+++ src/uavcan_node.cpp
@@ -130,7 +130,7 @@ void main(void *arg)
     uint8_t id = *(uint8_t *)arg;
 
     int res;
-    res = can.init(1000000);
+    res = can.init((uint32_t)1000000);
     if (res < 0) {
         node_fail("CAN init");
     }
